// apps/crm/client/src/components/tanstack-table/cards/companies/TableCompaniesOptionBtn.module.scss.d.ts
export type Styles = {
  companiesOptionBtn: string;
  companiesOptionBtn__menu: string;
  companiesOptionBtn__menuItem: string;
  companiesOptionBtn__menuItem__svg: string;
  companiesOptionBtn__menuItemWarning: string;
  companiesOptionBtn__menuItemWarning__svg: string;
  companiesOptionBtn__popover: string;
  companiesOptionBtn__svg: string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;